todos los consumibles separados (son 8) [revisar]

+++ entrega 2 grupo 2/entrega 01-grupo 2/Consumable.cs
@@ -8,7 +8,7 @@ namespace entrega_01_grupo_2
 {
     abstract class Consumable
     {
-        protected int Price;
+        protected double Price;
         protected bool Use;
         
         protected Consumable(int price)
